Extract night actions as NightPhase
Move runNightActions, revealNightSecrets, and attackIfNotGuarded into
NightPhase. PlayerManager.attack() is renamed to kill() and made public
for NightPhase to call. runTurn() now delegates to NightPhase.

Closes #74
