+++ src/main/kotlin/PlayerManager.kt
@@ -23,37 +23,9 @@ class PlayerManager(setup: GameSetup) {
         _nightDeath = null
     }
 
-    private fun runNightActions(nightNumber: Int) {
-        val decisions = _alivePlayers.map { it to it.buildNightAction(_alivePlayers, nightNumber == 1) }
-
-        val attacks = decisions.map { it.second }.filterIsInstance<NightAction.Attack>()
-        val guards = decisions.map { it.second }.filterIsInstance<NightAction.Guard>()
-        attackIfNotGuarded(attacks, guards)
-
-        revealNightSecrets(decisions)
-    }
-
-    private fun revealNightSecrets(decisions: List<Pair<Player, NightAction>>) {
-        decisions.forEach { (player, decision) ->
-            when (decision) {
-                is NightAction.None -> Unit
-                is NightAction.Attack -> Unit
-                is NightAction.Guard -> Unit
-                is NightAction.FirstNightDivine -> oracle.firstNightDivine(player, _alivePlayers)
-                is NightAction.Divine -> oracle.divine(player, decision.target)
-                is NightAction.MediumReveal -> oracle.mediumReveal(player, decision.target)
-            }
-        }
-    }
-
-    private fun attackIfNotGuarded(attacks: List<NightAction.Attack>, guards: List<NightAction.Guard>) {
-        val target = MajorityVoteResolver.resolve(attacks.map { it.target }) ?: return
-        if (guards.none { it.target === target }) attack(target)
-    }
-
     fun runTurn(nightNumber: Int) {
         GameEvent.TimeChanged.send(TimeOfDay.Night(nightNumber), allPlayers)
-        runNightActions(nightNumber)
+        NightPhase(this, oracle, nightNumber).proceed()
         GameEvent.TimeChanged.send(TimeOfDay.Morning, allPlayers)
         GameEvent.MorningReport.send(_nightDeath, allPlayers)
         bury()
@@ -81,7 +53,7 @@ class PlayerManager(setup: GameSetup) {
         die(mostVoted)
     }
 
-    private fun attack(target: Player) {
+    fun kill(target: Player) {
         GameEvent.PlayerAttacked.send(target, allPlayers)
         _nightDeath = target
         die(target)
